Factor out identifier data providers
