+++ tests/DataProvider/Identifier/XpathExpressionIdentifierDataProviderTrait.php
@@ -1,0 +1,48 @@
+<?php
+
+declare(strict_types=1);
+
+namespace webignition\BasilCompilableSourceFactory\Tests\DataProvider\Identifier;
+
+use webignition\BasilCompilableSourceFactory\Model\DomIdentifier;
+
+trait XpathExpressionIdentifierDataProviderTrait
+{
+    public function xpathExpressionIdentifierDataProvider(): array
+    {
+        return [
+            'xpath id selector' => [
+                'identifierString' => '$"//*[@id=\"element-id\"]"',
+                'expectedIdentifier' => new DomIdentifier('//*[@id="element-id"]'),
+            ],
+            'xpath attribute selector, position: null' => [
+                'identifierString' => '$"//input[@type=\"submit\"]"',
+                'expectedIdentifier' => new DomIdentifier('//input[@type="submit"]'),
+            ],
+            'xpath attribute selector; position: 1' => [
+                'identifierString' => '$"//input[@type=\"submit\"]":1',
+                'expectedIdentifier' => new DomIdentifier('//input[@type="submit"]', 1),
+            ],
+            'xpath attribute selector; position: 3' => [
+                'identifierString' => '$"//input[@type=\"submit\"]":3',
+                'expectedIdentifier' => new DomIdentifier('//input[@type="submit"]', 3),
+            ],
+            'xpath attribute selector; position: -1' => [
+                'identifierString' => '$"//input[@type=\"submit\"]":-1',
+                'expectedIdentifier' => new DomIdentifier('//input[@type="submit"]', -1),
+            ],
+            'xpath attribute selector; position: -3' => [
+                'identifierString' => '$"//input[@type=\"submit\"]":-3',
+                'expectedIdentifier' => new DomIdentifier('//input[@type="submit"]', -3),
+            ],
+            'xpath attribute selector; position: first' => [
+                'identifierString' => '$"//input[@type=\"submit\"]":first',
+                'expectedIdentifier' => new DomIdentifier('//input[@type="submit"]', 1),
+            ],
+            'xpath attribute selector; position: last' => [
+                'identifierString' => '$"//input[@type=\"submit\"]":last',
+                'expectedIdentifier' => new DomIdentifier('//input[@type="submit"]', -1),
+            ],
+        ];
+    }
+}
